FW/Meson: Rename sysdepds/win32 to sysdeps/windows
To match Meson target_machine.system()
