+++ framework/sysdeps/windows/cpu_affinity.cpp
