+++ framework/sysdeps/windows/fcntl.c
