+++ framework/sysdeps/windows/fcntl.h
