+++ framework/sysdeps/windows/sys/uio.h
