+++ framework/sysdeps/windows/thermal_monitor.hpp
