+++ framework/sysdeps/windows/ucrt-mapping.c
